Service Specification
I've tightened up the table showing the fields in a service description such that each field has an associated "priority" of REQUIRED, RECOMMENDED, or OPTIONAL (not used in this example).  This change illustrates my recommendation that the 1.0 spec needs to be specified more explicitly.  In addition, it carries out my recommendation to make the Prefetch field RECOMMENDED rather than required, so that each provider can decide whether to prefetch or allow the CDS service to fetch its own data as needed.