FHIR Resource Access
Incorporates all of the concepts we have discussed w.r.t. enabling an EHR to obtain a bearer token on behalf of a CDS service, and passing it on to that service in the service request.  I moved all of the pre-existing discussion re SMART into to introductory section so that the specification sections could be purely specification.  

I discovered that RFC7521 defines the "subject" parameter as "When using assertions as an authorization grant, the Subject identifies an authorized accessor for which the access token is being requested (typically, the resource owner or an authorized delegate)."  This seems to fit our purposes nicely, so I have used that as the identifier of the CDS Service that will be using the token.