Further clarify the purpose of the 'selections' context
It wasn't totally clear how or whether the CDS service was actually supposed to behave differently towards orders that were 'selected' vs. not.  The proposed wording is my guess at the intent, but happy to have the proposed wording changed to something else so long as it clarifies what the point is of differentiating selected from non-selected orders in terms of the behavior of the service.