ruleCode allows CDS Service to specify the rule used for card generation
Adding a ruleCode field to Card.source, of type FHIR Coding. This allows the CDS Service to expose metadata about the rule it used to generate the Card. Knowledge of this information can allow the CDS Client to be more nuanced in its presentation of the Card. 

Discussion: https://chat.fhir.org/#narrow/stream/179159-cds-hooks/topic/CDS.20Hooks.20feature.20request