rename `isRecommended`, make boolean, add'l guidance
`isRecommended` | OPTIONAL | *boolean* | Allows a service to indicate that a specific suggestion is recommended from all the available suggestions on the card. CDS Hooks clients may choose to influence their UI based on this value, such as pre-selecting, or highlighting recommended suggestions. Multiple suggestions MAY be recommended, if `card.selectionBehavior` is `any`. A CDS Service MUST only recommend one (or more) suggestions if that service expects the suggestion(s) to be accepted a majority of the time.